+++ src/pages/print/tokens.tsx
@@ -1,11 +1,11 @@
 import { type NextPage } from "next";
 import Head from "next/head";
-import Page from "~/components/Page";
-import Token from "~/components/Token";
+import Page from "~/components/print/Page";
+import Token from "~/components/print/Token";
 
 import roles from "~/data/roles/tb";
 
-const Home: NextPage = () => {
+const PrintTokenPage: NextPage = () => {
   return (
     <>
       <Head>
@@ -13,8 +13,8 @@ const Home: NextPage = () => {
         <meta name="description" content="Generated by create-t3-app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
+
       <main className="flex min-h-screen flex-col items-center justify-center bg-black">
-        {/* <div className="fixed top-0 left-0 h-[29.7cm] w-[21cm] bg-black/50"></div> */}
         <Page className="py-[0.5cm]">
           <div className="flex justify-around">
             <Token {...roles.chef} />
@@ -61,4 +61,4 @@ const Home: NextPage = () => {
   );
 };
 
-export default Home;
+export default PrintTokenPage;
